+++ frontend/package.json
@@ -18,6 +18,8 @@
     "react-dom": "^18.2.0",
     "react-icons": "^5.0.1",
     "react-router-dom": "^6.22.3",
+    "react-slick": "^0.30.2",
+    "slick-carousel": "^1.8.1",
     "swiper": "^11.0.7",
     "yup": "^1.4.0"
   },
